allow inputs_embeds to be passed to forward call
In the forward call of the model, the inputs_embeds
was being set by the internal helper call
self.prepare_inputs_labels_for_multimodal().
However, this is also a valid parameter and
should probably not be overwritten in all cases.

This workaround only assigns this value internally
when the value provided still has the default
value of None. Otherwise, the value passed to
the call is honored.

Was tested on code calling the model via
`model(inputs_embeds=embeds)` and seemed
to work as expected.