Update the link for the SEED-Bench benchmark instructions
The DATASET.md file of Seed-Bench now contains instructions for both versions of Seed-Bench (v1 and v2). In the LLaVA-1.5 paper, you reference the paper presenting Seed-Bench-1. So, this link as it is now might be confusing, and I'd like to add an anchor pointing to the section about data preparation for Seed-Bench-1.